@@ -13,14 +13,23 @@ if PROJECT_ROOT not in sys.path:
 
 import json
 from utils.app_logging import get_logger
-from typing import Dict, Any
+from typing import Dict, Any, cast
 from datetime import datetime, timedelta
+from pathlib import Path
 
 from paths import get_raw_news_dir
 
 logger = get_logger(__name__)
 
-def load_article(article_id: str, max_days: int = 30) -> Dict[str, Any] | None:
+def _load_json_object(path: Path) -> dict[str, Any]:
+    """Load JSON and guarantee the top-level is an object (dict)."""
+    with path.open("r", encoding="utf-8") as f:
+        obj = json.load(f)
+    if not isinstance(obj, dict):
+        raise TypeError(f"Expected JSON object at top level in {path}, got {type(obj).__name__}")
+    return cast(dict[str, Any], obj)
+
+def load_article(article_id: str, max_days: int = 30) -> Dict[str, str] | None:
     """
     Loads a single article from raw news directories by its unique ID.
     Tries today, then yesterday, etc., up to max_days back.
@@ -39,28 +48,31 @@ def load_article(article_id: str, max_days: int = 30) -> Dict[str, Any] | None:
         day = (today - timedelta(days=days_back)).strftime("%Y-%m-%d")
         day_dir = get_raw_news_dir(day)
         article_path = day_dir / f"{article_id}.json"
-        logger.debug(f"Trying to load article {article_id} from {article_path}")
-        if article_path.exists():
-            try:
-                with open(article_path, "r") as f:
-                    obj = json.load(f)
-                logger.debug(f"Loaded article {article_id} from {article_path}")
-                if isinstance(obj, dict) and "data" in obj and isinstance(obj["data"], dict):
-                    return obj["data"]
-                return obj
-            except Exception as e:
-                logger.error(f"Failed to load article {article_id} from {article_path}: {e}")
-                continue
-        else:
-            logger.debug(f"Article {article_id} not found in {article_path}")
-    logger.error(f"Article {article_id} not found in any raw_news day directory (checked {max_days} days)")
-    raise FileNotFoundError(f"Article {article_id} not found in any raw_news day directory")
+        logger.debug("Trying to load article %s from %s", article_id, article_path)
 
-if __name__ == "__main__":
-    test_id = "5JO4Z4OKV"
-    logger.info(f"Testing load_article with id: {test_id}")
-    try:
-        article = load_article(test_id)
-        logger.info("Loaded article: %s", article)
-    except Exception as e:
-        logger.error(f"Error: {e}")
+        if not article_path.exists():
+            logger.debug("Article %s not found in %s", article_id, article_path)
+            continue
+
+        try:
+            obj = _load_json_object(article_path)
+            logger.debug("Loaded article %s from %s", article_id, article_path)
+
+            # Handle legacy wrapper
+            inner = obj.get("data")
+            if isinstance(inner, dict):
+                return cast(Dict[str, str], inner)
+            return obj  # guaranteed dict[str, Any]
+
+        except (json.JSONDecodeError, TypeError) as e:
+            logger.error("Failed to parse %s: %s", article_path, e)
+            continue
+        except Exception as e:
+            logger.exception("Unexpected error reading %s: %s", article_path, e)
+            continue
+
+    logger.error(
+        "Article %s not found in any raw_news day directory (checked %d days)",
+        article_id, max_days,
+    )
+    raise FileNotFoundError(f"Article {article_id} not found")
