@@ -1,15 +1,7 @@
-import sys, os
+import os
 import datetime
 from fpdf import FPDF
-
-# --- Canonical project root import pattern ---
-PROJECT_ROOT = os.path.dirname(os.path.abspath(__file__))
-while not os.path.exists(os.path.join(PROJECT_ROOT, "main.py")) and PROJECT_ROOT != "/":
-    PROJECT_ROOT = os.path.dirname(PROJECT_ROOT)
-if PROJECT_ROOT not in sys.path:
-    sys.path.insert(0, PROJECT_ROOT)
-
-from src.graph.ops.get_topic_id_by_name import get_topic_id_by_name
+from src.graph.ops.topic import get_topic_id_by_name
 from src.analysis.utils.report_aggregator import aggregate_reports
 from utils import app_logging
 
